[RFC] Fix file not found error on Windows (#24)
When using Intellij IDEA to edit a project, Windows users see a file not found error on L247. I believe this is due to how Intellij saves files (through renames instead of direct writes). So when bra tries to get the modified time for the file the file isn't there any more.

I tested this change on Windows and on Linux and everything seemed fine. The RFC is calling for others to test this change locally.